chore: separate schemas from get my jobs hooks
