+++ packages/apps/human-app/frontend/src/modules/worker/jobs/hooks/index.ts
@@ -4,3 +4,4 @@ export * from './use-job-types-oracles-table';
 export * from './use-jobs-filter-store';
 export * from './use-my-jobs-filter-store';
 export * from './use-get-ui-config';
+export * from './use-get-my-jobs-data';
