+++ packages/apps/human-app/frontend/src/modules/worker/jobs/my-jobs/components/desktop/columns.tsx
@@ -10,11 +10,11 @@ import { Chip } from '@/shared/components/ui/chip';
 import type { JobType } from '@/modules/smart-contracts/EthKVStore/config';
 import { formatDate } from '@/shared/helpers/date';
 import {
-  type MyJob,
   EvmAddress,
   RewardAmount,
   MyJobsTableActions,
 } from '../../../components';
+import { type MyJob } from '../../schemas';
 import { StatusChip } from './status-chip';
 import { MyJobsExpiresAtSort } from './my-jobs-expires-at-sort';
 import { MyJobsJobTypeFilter } from './my-jobs-job-type-filter';
